Fixing netstandard build of System.Collection.Imutable.Tests project
