+++ src/System.Collections.Immutable/tests/ImmutableDictionaryBuilderTestBase.cs
@@ -8,7 +8,7 @@ using Xunit;
 
 namespace System.Collections.Immutable.Tests
 {
-    public abstract class ImmutableDictionaryBuilderTestBase : ImmutablesTestBase
+    public abstract partial class ImmutableDictionaryBuilderTestBase : ImmutablesTestBase
     {
         [Fact]
         public void Add()
@@ -142,22 +142,6 @@ namespace System.Collections.Immutable.Tests
             Assert.True(builder.TryGetValue("six", out value) && value == 6);
             Assert.False(builder.TryGetValue("four", out value));
             Assert.Equal(0, value);
-        }
-
-        [Fact]
-        public void TryGetKey()
-        {
-            var builder = Empty<int>(StringComparer.OrdinalIgnoreCase)
-                .Add("a", 1).ToBuilder();
-            string actualKey;
-            Assert.True(TryGetKeyHelper(builder, "a", out actualKey));
-            Assert.Equal("a", actualKey);
-
-            Assert.True(TryGetKeyHelper(builder, "A", out actualKey));
-            Assert.Equal("a", actualKey);
-
-            Assert.False(TryGetKeyHelper(builder, "b", out actualKey));
-            Assert.Equal("b", actualKey);
         }
 
         [Fact]
