+++ src/System.Collections.Immutable/tests/System.Collections.Immutable.Tests.csproj
@@ -11,7 +11,6 @@
     </Compile>
     <Compile Include="BadHasher.cs" />
     <Compile Include="EverythingEqual.cs" />
-    <Compile Include="ImmutableArrayBuilderTest.cs" />
     <Compile Include="ImmutableArrayExtensionsTest.cs" />
     <Compile Include="ImmutableArrayTest.cs" />
     <Compile Include="ImmutableArrayTestBase.cs" />
@@ -21,23 +20,18 @@
     <Compile Include="ImmutableDictionaryBuilderTest.cs" />
     <Compile Include="ImmutableHashSetBuilderTest.cs" />
     <Compile Include="ImmutableInterlockedTests.cs" />
-    <Compile Include="ImmutableListTestBase.cs" />
-    <Compile Include="ImmutableListBuilderTest.cs" />
     <Compile Include="ImmutableDictionaryTest.cs" />
     <Compile Include="ImmutableHashSetTest.cs" />
-    <Compile Include="ImmutableListTest.cs" />
     <Compile Include="ImmutableDictionaryTestBase.cs" />
     <Compile Include="ImmutableQueueTest.cs" />
     <Compile Include="ImmutableSetTest.cs" />
     <Compile Include="ImmutableSortedDictionaryBuilderTest.cs" />
     <Compile Include="ImmutableSortedDictionaryTest.cs" />
-    <Compile Include="ImmutableSortedSetBuilderDebuggerProxyTest.cs" />
     <Compile Include="ImmutableSortedSetBuilderTest.cs" />
     <Compile Include="ImmutableSortedSetTest.cs" />
     <Compile Include="ImmutableStackTest.cs" />
     <Compile Include="ImmutableTestBase.cs" />
     <Compile Include="IndexOfTests.cs" />
-    <Compile Include="RequiresTests.cs" />
     <Compile Include="SimpleElementImmutablesTestBase.cs" />
     <Compile Include="TestExtensionsMethods.cs" />
     <Compile Include="$(CommonTestPath)\System\Diagnostics\DebuggerAttributes.cs">
@@ -103,10 +97,25 @@
     <None Include="ClassDiagram1.cd" />
     <Compile Include="ImmutableArray\ImmutableArray.Generic.Tests.cs" />
   </ItemGroup>
-  <ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' != 'netstandard'">
     <!-- Some internal types are needed, so we reference the implementation assembly, rather than the reference assembly. -->
     <TargetingPackExclusions Include="System.Collections.Immutable" />
     <ReferenceFromRuntime Include="System.Collections.Immutable" />
+    <Compile Include="ImmutableListTestBase.cs" />
+    <Compile Include="ImmutableListTest.cs" />
+    <Compile Include="ImmutableListBuilderTest.cs" />
+    <Compile Include="RequiresTests.cs" />
+    <Compile Include="ImmutableArrayBuilderTest.cs" />
+    <Compile Include="ImmutableDictionaryTestBase.nonnetstandard.cs" />
+    <Compile Include="ImmutableDictionaryTest.nonnetstandard.cs" />
+    <Compile Include="TestExtensionsMethods.nonnetstandard.cs" />
+    <Compile Include="ImmutableSetTest.nonnetstandard.cs" />
+    <Compile Include="ImmutableHashSetTest.nonnetstandard.cs" />
+    <Compile Include="ImmutableSortedSetTest.nonnetstandard.cs" />
+    <Compile Include="ImmutableTestBase.nonnetstandard.cs" />
+    <Compile Include="ImmutableDictionaryBuilderTestBase.nonnetstandard.cs" />
+    <Compile Include="ImmutableSortedSetBuilderDebuggerProxyTest.cs" />
+    <Compile Include="ImmutableSortedSetBuilderTest.nonnetstandard.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\$(AssemblyName).rd.xml" />
